Converted Presenter/Collector into a singular Collector interface implementsed by Decoder/Encoder which consolidates the load and save methods. Encoder.collect will save the data presented, Decoder.collect will return the data in the save file. Introduced Stations which will be an extra layer of complexity/organization, an example is only SpaceShip's assigned to RepairStation's will be repaired instead of the all the ships in a hexagon being repaired at the same time.

+++ RemWfSCore/src/com/rem/wfs/environment/Statistic.java
@@ -3,5 +3,5 @@ package com.rem.wfs.environment;
 public interface Statistic {
 
 	public float getValue();
-	public float getLimit();
+	public int getLimit();
 }
